requirements.txt
pytest==7.4.3
pydantic==2.5.0
python-dotenv==1.0.0
PyMuPDF==1.24.1
pdf2image==1.17.0
Pillow==10.3.0
pytesseract==0.3.10
reportlab==4.0.7
rapidfuzz==3.6.1
PyPDF2==3.0.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
python-multipart==0.0.9
